Refactoring folders and sources, and remove version 1
