--- js/test/dummy.test.js
@@ -1,3 +1,0 @@
-test("dummy test", () => {
-    expect(0).toBe(0);
-})
