+++ src/drawPaddle.js
