+++ src/drawStage.js
