--- src/levels.js
@@ -1,29 +1,0 @@
-let colors = ["#edbf74", "#6eaf99", "#e68b86", "#f9a6c2", "#fed070","#b26e47", "#88adb3", "#94f2f4", "#e8cfc8", "#d4a3a9", "#ecc3c1", "#edac8c", "#c9b4bb", "#f5deb2", "#c1cf94", "#e38c95", "#348f90", "#afebe1", "#5d97af", "#ffbfa1", "#e2895f", "#9daadf", "#ffabb8", "#bd84a4", "#9898bc", "#487ba6", "#adddcf", "#6eaf99", "#e7b74b", "#f2a557", "#cb7d8b", "#88c0e1", "#f0c4d5", "#f7c9b1", "#eab251", "#9f6b53" ]
-
-
-
-let levels = {1: [{xInd:9,yInd: 3, color:colors[0]},
-                    {xInd:9,yInd: 4, color:colors[1]},
-                    {xInd:9,yInd: 5, color:colors[2]},
-                    {xInd:9,yInd: 6, color:colors[0]},
-                    {xInd:9,yInd: 7, color:colors[1]},
-                    {xInd:9,yInd: 8, color:colors[2]},
-                    {xInd:9,yInd: 9, color:colors[0]},
-                    {xInd:9,yInd: 10, color:colors[1]},
-                    {xInd:9,yInd: 11, color:colors[2]},
-                    {xInd:9,yInd: 12, color:colors[0]},
-                    {xInd:9,yInd: 13, color:colors[1]},
-                    {xInd:9,yInd: 14, color:colors[2]}],
-                2: [{xInd: 6, yInd: 4, color: '#9daadf'},{xInd: 6, yInd: 6, color: '#9daadf'},{xInd: 6, yInd: 10, color: '#9daadf'},{xInd: 6, yInd: 12, color: '#9daadf'},{xInd: 7, yInd: 3, color: '#f9a6c2'},{xInd: 7, yInd: 7, color: '#f9a6c2'},{xInd: 7, yInd: 9, color: '#f9a6c2'},{xInd: 7, yInd: 13, color: '#f9a6c2'},{xInd: 8, yInd: 4, color: '#edbf74'},{xInd: 8, yInd: 6, color: '#edbf74'},{xInd: 8, yInd: 10, color: '#edbf74'},{xInd: 8, yInd: 12, color: '#edbf74'},{xInd: 9, yInd: 3, color: '#c1cf94'},{xInd: 9, yInd: 7, color: '#c1cf94'},{xInd: 9, yInd: 9, color: '#c1cf94'},{xInd: 9, yInd: 13, color: '#c1cf94'},{xInd: 10, yInd: 4, color: '#edbf74'},{xInd: 10, yInd: 6, color: '#edbf74'},{xInd: 10, yInd: 10, color: '#edbf74'},{xInd: 10, yInd: 12, color: '#edbf74'},{xInd: 11, yInd: 3, color: '#f9a6c2'},{xInd: 11, yInd: 7, color: '#f9a6c2'},{xInd: 11, yInd: 9, color: '#f9a6c2'},{xInd: 11, yInd: 13, color: '#f9a6c2'},{xInd: 12, yInd: 4, color: '#9daadf'},{xInd: 12, yInd: 6, color: '#9daadf'},{xInd: 12, yInd: 10, color: '#9daadf'},{xInd: 12, yInd: 12, color: '#9daadf'},],  
-                3: [{xInd: 3, yInd: 6, color: '#9daadf'},{xInd: 3, yInd: 8, color: '#cb7d8b'},{xInd: 3, yInd: 10, color: '#9daadf'},{xInd: 3, yInd: 12, color: '#cb7d8b'},{xInd: 4, yInd: 6, color: '#cb7d8b'},{xInd: 4, yInd: 8, color: '#9daadf'},{xInd: 4, yInd: 10, color: '#cb7d8b'},{xInd: 4, yInd: 12, color: '#9daadf'},{xInd: 5, yInd: 6, color: '#9daadf'},{xInd: 5, yInd: 8, color: '#cb7d8b'},{xInd: 5, yInd: 10, color: '#9daadf'},{xInd: 5, yInd: 12, color: '#cb7d8b'},{xInd: 6, yInd: 6, color: '#cb7d8b'},{xInd: 6, yInd: 8, color: '#9daadf'},{xInd: 6, yInd: 10, color: '#cb7d8b'},{xInd: 6, yInd: 12, color: '#9daadf'},{xInd: 7, yInd: 6, color: '#9daadf'},{xInd: 7, yInd: 8, color: '#cb7d8b'},{xInd: 7, yInd: 10, color: '#9daadf'},{xInd: 7, yInd: 12, color: '#cb7d8b'},{xInd: 8, yInd: 6, color: '#cb7d8b'},{xInd: 8, yInd: 8, color: '#9daadf'},{xInd: 8, yInd: 10, color: '#cb7d8b'},{xInd: 8, yInd: 12, color: '#9daadf'},{xInd: 9, yInd: 6, color: '#9daadf'},{xInd: 9, yInd: 8, color: '#cb7d8b'},{xInd: 9, yInd: 10, color: '#9daadf'},{xInd: 9, yInd: 12, color: '#cb7d8b'},{xInd: 10, yInd: 6, color: '#cb7d8b'},{xInd: 10, yInd: 8, color: '#9daadf'},{xInd: 10, yInd: 10, color: '#cb7d8b'},{xInd: 10, yInd: 12, color: '#9daadf'},{xInd: 11, yInd: 6, color: '#9daadf'},{xInd: 11, yInd: 8, color: '#cb7d8b'},{xInd: 11, yInd: 10, color: '#9daadf'},{xInd: 11, yInd: 12, color: '#cb7d8b'},{xInd: 12, yInd: 6, color: '#cb7d8b'},{xInd: 12, yInd: 8, color: '#9daadf'},{xInd: 12, yInd: 10, color: '#cb7d8b'},{xInd: 12, yInd: 12, color: '#9daadf'},{xInd: 13, yInd: 6, color: '#9daadf'},{xInd: 13, yInd: 8, color: '#cb7d8b'},{xInd: 13, yInd: 10, color: '#9daadf'},{xInd: 13, yInd: 12, color: '#cb7d8b'},{xInd: 14, yInd: 6, color: '#cb7d8b'},{xInd: 14, yInd: 8, color: '#9daadf'},{xInd: 14, yInd: 10, color: '#cb7d8b'},{xInd: 14, yInd: 12, color: '#9daadf'},{xInd: 15, yInd: 6, color: '#9daadf'},{xInd: 15, yInd: 8, color: '#cb7d8b'},{xInd: 15, yInd: 10, color: '#9daadf'},{xInd: 15, yInd: 12, color: '#cb7d8b'},], 
-                4: [{xInd: 0, yInd: 24, color: '#adddcf'},{xInd: 0, yInd: 25, color: '#e68b86'},{xInd: 0, yInd: 26, color: '#edbf74'},{xInd: 0, yInd: 27, color: '#5d97af'},{xInd: 0, yInd: 28, color: '#9898bc'},{xInd: 1, yInd: 25, color: '#e68b86'},{xInd: 1, yInd: 26, color: '#edbf74'},{xInd: 1, yInd: 27, color: '#5d97af'},{xInd: 1, yInd: 28, color: '#9898bc'},{xInd: 2, yInd: 26, color: '#edbf74'},{xInd: 2, yInd: 27, color: '#5d97af'},{xInd: 2, yInd: 28, color: '#9898bc'},{xInd: 3, yInd: 27, color: '#5d97af'},{xInd: 3, yInd: 28, color: '#9898bc'},{xInd: 4, yInd: 28, color: '#9898bc'},{xInd: 14, yInd: 28, color: '#9898bc'},{xInd: 15, yInd: 27, color: '#5d97af'},{xInd: 15, yInd: 28, color: '#9898bc'},{xInd: 16, yInd: 26, color: '#edbf74'},{xInd: 16, yInd: 27, color: '#5d97af'},{xInd: 16, yInd: 28, color: '#9898bc'},{xInd: 17, yInd: 25, color: '#e68b86'},{xInd: 17, yInd: 26, color: '#edbf74'},{xInd: 17, yInd: 27, color: '#5d97af'},{xInd: 17, yInd: 28, color: '#9898bc'},{xInd: 18, yInd: 24, color: '#adddcf'},{xInd: 18, yInd: 25, color: '#e68b86'},{xInd: 18, yInd: 26, color: '#edbf74'},{xInd: 18, yInd: 27, color: '#5d97af'},{xInd: 18, yInd: 28, color: '#9898bc'},],
-                5: [{xInd: 3, yInd: 2, color: '#ffabb8'},{xInd: 3, yInd: 3, color: '#ffabb8'},{xInd: 3, yInd: 4, color: '#ffabb8'},{xInd: 3, yInd: 5, color: '#ffabb8'},{xInd: 3, yInd: 6, color: '#ffabb8'},{xInd: 3, yInd: 7, color: '#ffabb8'},{xInd: 3, yInd: 8, color: '#ffabb8'},{xInd: 3, yInd: 9, color: '#ffabb8'},{xInd: 3, yInd: 10, color: '#ffabb8'},{xInd: 3, yInd: 11, color: '#ffabb8'},{xInd: 3, yInd: 12, color: '#ffabb8'},{xInd: 3, yInd: 13, color: '#ffabb8'},{xInd: 3, yInd: 14, color: '#ffabb8'},{xInd: 3, yInd: 15, color: '#ffabb8'},{xInd: 3, yInd: 16, color: '#ffabb8'},{xInd: 3, yInd: 17, color: '#ffabb8'},{xInd: 3, yInd: 18, color: '#ffabb8'},{xInd: 3, yInd: 19, color: '#ffabb8'},{xInd: 4, yInd: 2, color: '#ffabb8'},{xInd: 5, yInd: 2, color: '#ffabb8'},{xInd: 5, yInd: 4, color: '#88c0e1'},{xInd: 5, yInd: 5, color: '#88c0e1'},{xInd: 5, yInd: 6, color: '#88c0e1'},{xInd: 5, yInd: 7, color: '#88c0e1'},{xInd: 5, yInd: 8, color: '#88c0e1'},{xInd: 5, yInd: 9, color: '#88c0e1'},{xInd: 5, yInd: 10, color: '#88c0e1'},{xInd: 5, yInd: 11, color: '#88c0e1'},{xInd: 5, yInd: 12, color: '#88c0e1'},{xInd: 5, yInd: 13, color: '#88c0e1'},{xInd: 5, yInd: 14, color: '#88c0e1'},{xInd: 5, yInd: 15, color: '#88c0e1'},{xInd: 5, yInd: 16, color: '#88c0e1'},{xInd: 5, yInd: 17, color: '#88c0e1'},{xInd: 5, yInd: 18, color: '#88c0e1'},{xInd: 6, yInd: 2, color: '#ffabb8'},{xInd: 6, yInd: 4, color: '#88c0e1'},{xInd: 7, yInd: 2, color: '#ffabb8'},{xInd: 7, yInd: 4, color: '#88c0e1'},{xInd: 8, yInd: 2, color: '#ffabb8'},{xInd: 8, yInd: 4, color: '#88c0e1'},{xInd: 9, yInd: 2, color: '#ffabb8'},{xInd: 9, yInd: 4, color: '#88c0e1'},{xInd: 10, yInd: 2, color: '#ffabb8'},{xInd: 10, yInd: 4, color: '#88c0e1'},{xInd: 11, yInd: 2, color: '#ffabb8'},{xInd: 11, yInd: 4, color: '#88c0e1'},{xInd: 12, yInd: 2, color: '#ffabb8'},{xInd: 12, yInd: 4, color: '#88c0e1'},{xInd: 13, yInd: 2, color: '#ffabb8'},{xInd: 13, yInd: 4, color: '#88c0e1'},{xInd: 13, yInd: 5, color: '#88c0e1'},{xInd: 13, yInd: 6, color: '#88c0e1'},{xInd: 13, yInd: 7, color: '#88c0e1'},{xInd: 13, yInd: 8, color: '#88c0e1'},{xInd: 13, yInd: 9, color: '#88c0e1'},{xInd: 13, yInd: 10, color: '#88c0e1'},{xInd: 13, yInd: 11, color: '#88c0e1'},{xInd: 13, yInd: 12, color: '#88c0e1'},{xInd: 13, yInd: 13, color: '#88c0e1'},{xInd: 13, yInd: 14, color: '#88c0e1'},{xInd: 13, yInd: 15, color: '#88c0e1'},{xInd: 13, yInd: 16, color: '#88c0e1'},{xInd: 13, yInd: 17, color: '#88c0e1'},{xInd: 13, yInd: 18, color: '#88c0e1'},{xInd: 14, yInd: 2, color: '#ffabb8'},{xInd: 15, yInd: 2, color: '#ffabb8'},{xInd: 15, yInd: 3, color: '#ffabb8'},{xInd: 15, yInd: 4, color: '#ffabb8'},{xInd: 15, yInd: 5, color: '#ffabb8'},{xInd: 15, yInd: 6, color: '#ffabb8'},{xInd: 15, yInd: 7, color: '#ffabb8'},{xInd: 15, yInd: 8, color: '#ffabb8'},{xInd: 15, yInd: 9, color: '#ffabb8'},{xInd: 15, yInd: 10, color: '#ffabb8'},{xInd: 15, yInd: 11, color: '#ffabb8'},{xInd: 15, yInd: 12, color: '#ffabb8'},{xInd: 15, yInd: 13, color: '#ffabb8'},{xInd: 15, yInd: 14, color: '#ffabb8'},{xInd: 15, yInd: 15, color: '#ffabb8'},{xInd: 15, yInd: 16, color: '#ffabb8'},{xInd: 15, yInd: 17, color: '#ffabb8'},{xInd: 15, yInd: 18, color: '#ffabb8'},{xInd: 15, yInd: 19, color: '#ffabb8'},],
-                6: [{xInd: 2, yInd: 11, color: '#adddcf'},{xInd: 3, yInd: 10, color: '#adddcf'},{xInd: 3, yInd: 11, color: '#ecc3c1'},{xInd: 3, yInd: 12, color: '#adddcf'},{xInd: 4, yInd: 9, color: '#adddcf'},{xInd: 4, yInd: 10, color: '#ecc3c1'},{xInd: 4, yInd: 11, color: '#adddcf'},{xInd: 4, yInd: 12, color: '#ecc3c1'},{xInd: 4, yInd: 13, color: '#adddcf'},{xInd: 5, yInd: 8, color: '#adddcf'},{xInd: 5, yInd: 9, color: '#ecc3c1'},{xInd: 5, yInd: 10, color: '#adddcf'},{xInd: 5, yInd: 11, color: '#ecc3c1'},{xInd: 5, yInd: 12, color: '#adddcf'},{xInd: 5, yInd: 13, color: '#ecc3c1'},{xInd: 5, yInd: 14, color: '#afebe1'},{xInd: 6, yInd: 7, color: '#adddcf'},{xInd: 6, yInd: 8, color: '#ecc3c1'},{xInd: 6, yInd: 9, color: '#adddcf'},{xInd: 6, yInd: 10, color: '#ecc3c1'},{xInd: 6, yInd: 11, color: '#adddcf'},{xInd: 6, yInd: 12, color: '#ecc3c1'},{xInd: 6, yInd: 13, color: '#adddcf'},{xInd: 6, yInd: 14, color: '#ecc3c1'},{xInd: 6, yInd: 15, color: '#adddcf'},{xInd: 7, yInd: 6, color: '#adddcf'},{xInd: 7, yInd: 7, color: '#ecc3c1'},{xInd: 7, yInd: 8, color: '#adddcf'},{xInd: 7, yInd: 9, color: '#ecc3c1'},{xInd: 7, yInd: 10, color: '#adddcf'},{xInd: 7, yInd: 11, color: '#ecc3c1'},{xInd: 7, yInd: 12, color: '#adddcf'},{xInd: 7, yInd: 13, color: '#ecc3c1'},{xInd: 7, yInd: 14, color: '#adddcf'},{xInd: 7, yInd: 15, color: '#ecc3c1'},{xInd: 7, yInd: 16, color: '#adddcf'},{xInd: 8, yInd: 5, color: '#adddcf'},{xInd: 8, yInd: 6, color: '#ecc3c1'},{xInd: 8, yInd: 7, color: '#adddcf'},{xInd: 8, yInd: 8, color: '#ecc3c1'},{xInd: 8, yInd: 9, color: '#adddcf'},{xInd: 8, yInd: 10, color: '#ecc3c1'},{xInd: 8, yInd: 11, color: '#adddcf'},{xInd: 8, yInd: 12, color: '#ecc3c1'},{xInd: 8, yInd: 13, color: '#adddcf'},{xInd: 8, yInd: 14, color: '#ecc3c1'},{xInd: 8, yInd: 15, color: '#adddcf'},{xInd: 8, yInd: 16, color: '#ecc3c1'},{xInd: 8, yInd: 17, color: '#adddcf'},{xInd: 9, yInd: 4, color: '#adddcf'},{xInd: 9, yInd: 5, color: '#ecc3c1'},{xInd: 9, yInd: 6, color: '#adddcf'},{xInd: 9, yInd: 7, color: '#ecc3c1'},{xInd: 9, yInd: 8, color: '#adddcf'},{xInd: 9, yInd: 9, color: '#ecc3c1'},{xInd: 9, yInd: 10, color: '#adddcf'},{xInd: 9, yInd: 11, color: '#ecc3c1'},{xInd: 9, yInd: 12, color: '#adddcf'},{xInd: 9, yInd: 13, color: '#ecc3c1'},{xInd: 9, yInd: 14, color: '#adddcf'},{xInd: 9, yInd: 15, color: '#ecc3c1'},{xInd: 9, yInd: 16, color: '#adddcf'},{xInd: 9, yInd: 17, color: '#ecc3c1'},{xInd: 9, yInd: 18, color: '#adddcf'},{xInd: 10, yInd: 5, color: '#adddcf'},{xInd: 10, yInd: 6, color: '#ecc3c1'},{xInd: 10, yInd: 7, color: '#adddcf'},{xInd: 10, yInd: 8, color: '#ecc3c1'},{xInd: 10, yInd: 9, color: '#adddcf'},{xInd: 10, yInd: 10, color: '#ecc3c1'},{xInd: 10, yInd: 11, color: '#adddcf'},{xInd: 10, yInd: 12, color: '#ecc3c1'},{xInd: 10, yInd: 13, color: '#adddcf'},{xInd: 10, yInd: 14, color: '#ecc3c1'},{xInd: 10, yInd: 15, color: '#adddcf'},{xInd: 10, yInd: 16, color: '#ecc3c1'},{xInd: 10, yInd: 17, color: '#adddcf'},{xInd: 11, yInd: 6, color: '#adddcf'},{xInd: 11, yInd: 7, color: '#ecc3c1'},{xInd: 11, yInd: 8, color: '#adddcf'},{xInd: 11, yInd: 9, color: '#ecc3c1'},{xInd: 11, yInd: 10, color: '#adddcf'},{xInd: 11, yInd: 11, color: '#ecc3c1'},{xInd: 11, yInd: 12, color: '#adddcf'},{xInd: 11, yInd: 13, color: '#ecc3c1'},{xInd: 11, yInd: 14, color: '#adddcf'},{xInd: 11, yInd: 15, color: '#ecc3c1'},{xInd: 11, yInd: 16, color: '#adddcf'},{xInd: 12, yInd: 7, color: '#adddcf'},{xInd: 12, yInd: 8, color: '#ecc3c1'},{xInd: 12, yInd: 9, color: '#adddcf'},{xInd: 12, yInd: 10, color: '#ecc3c1'},{xInd: 12, yInd: 11, color: '#adddcf'},{xInd: 12, yInd: 12, color: '#ecc3c1'},{xInd: 12, yInd: 13, color: '#adddcf'},{xInd: 12, yInd: 14, color: '#ecc3c1'},{xInd: 12, yInd: 15, color: '#adddcf'},{xInd: 13, yInd: 8, color: '#adddcf'},{xInd: 13, yInd: 9, color: '#ecc3c1'},{xInd: 13, yInd: 10, color: '#adddcf'},{xInd: 13, yInd: 11, color: '#ecc3c1'},{xInd: 13, yInd: 12, color: '#adddcf'},{xInd: 13, yInd: 13, color: '#ecc3c1'},{xInd: 13, yInd: 14, color: '#adddcf'},{xInd: 14, yInd: 9, color: '#adddcf'},{xInd: 14, yInd: 10, color: '#ecc3c1'},{xInd: 14, yInd: 11, color: '#adddcf'},{xInd: 14, yInd: 12, color: '#ecc3c1'},{xInd: 14, yInd: 13, color: '#adddcf'},{xInd: 15, yInd: 10, color: '#adddcf'},{xInd: 15, yInd: 11, color: '#ecc3c1'},{xInd: 15, yInd: 12, color: '#adddcf'},{xInd: 16, yInd: 11, color: '#adddcf'},],
-                7: [{xInd: 1, yInd: 1, color: '#5d97af'},{xInd: 1, yInd: 2, color: '#5d97af'},{xInd: 1, yInd: 27, color: '#f9a6c2'},{xInd: 1, yInd: 28, color: '#f9a6c2'},{xInd: 2, yInd: 1, color: '#5d97af'},{xInd: 2, yInd: 2, color: '#5d97af'},{xInd: 2, yInd: 27, color: '#f9a6c2'},{xInd: 2, yInd: 28, color: '#f9a6c2'},{xInd: 3, yInd: 3, color: '#5d97af'},{xInd: 3, yInd: 4, color: '#5d97af'},{xInd: 3, yInd: 25, color: '#f9a6c2'},{xInd: 3, yInd: 26, color: '#f9a6c2'},{xInd: 4, yInd: 3, color: '#5d97af'},{xInd: 4, yInd: 4, color: '#5d97af'},{xInd: 4, yInd: 25, color: '#f9a6c2'},{xInd: 4, yInd: 26, color: '#f9a6c2'},{xInd: 5, yInd: 5, color: '#5d97af'},{xInd: 5, yInd: 6, color: '#5d97af'},{xInd: 5, yInd: 23, color: '#f9a6c2'},{xInd: 5, yInd: 24, color: '#f9a6c2'},{xInd: 6, yInd: 5, color: '#5d97af'},{xInd: 6, yInd: 6, color: '#5d97af'},{xInd: 6, yInd: 23, color: '#f9a6c2'},{xInd: 6, yInd: 24, color: '#f9a6c2'},{xInd: 7, yInd: 7, color: '#5d97af'},{xInd: 7, yInd: 8, color: '#5d97af'},{xInd: 7, yInd: 21, color: '#f9a6c2'},{xInd: 7, yInd: 22, color: '#f9a6c2'},{xInd: 8, yInd: 7, color: '#5d97af'},{xInd: 8, yInd: 8, color: '#5d97af'},{xInd: 8, yInd: 9, color: '#c1cf94'},{xInd: 8, yInd: 10, color: '#c1cf94'},{xInd: 8, yInd: 11, color: '#c1cf94'},{xInd: 8, yInd: 12, color: '#c1cf94'},{xInd: 8, yInd: 13, color: '#c1cf94'},{xInd: 8, yInd: 14, color: '#c1cf94'},{xInd: 8, yInd: 15, color: '#c1cf94'},{xInd: 8, yInd: 16, color: '#c1cf94'},{xInd: 8, yInd: 17, color: '#c1cf94'},{xInd: 8, yInd: 18, color: '#c1cf94'},{xInd: 8, yInd: 19, color: '#c1cf94'},{xInd: 8, yInd: 20, color: '#c1cf94'},{xInd: 8, yInd: 21, color: '#f9a6c2'},{xInd: 8, yInd: 22, color: '#f9a6c2'},{xInd: 9, yInd: 9, color: '#c1cf94'},{xInd: 9, yInd: 10, color: '#c1cf94'},{xInd: 9, yInd: 11, color: '#c1cf94'},{xInd: 9, yInd: 12, color: '#c1cf94'},{xInd: 9, yInd: 13, color: '#c1cf94'},{xInd: 9, yInd: 14, color: '#c1cf94'},{xInd: 9, yInd: 15, color: '#c1cf94'},{xInd: 9, yInd: 16, color: '#c1cf94'},{xInd: 9, yInd: 17, color: '#c1cf94'},{xInd: 9, yInd: 18, color: '#c1cf94'},{xInd: 9, yInd: 19, color: '#c1cf94'},{xInd: 9, yInd: 20, color: '#c1cf94'},{xInd: 10, yInd: 7, color: '#fed070'},{xInd: 10, yInd: 8, color: '#fed070'},{xInd: 10, yInd: 9, color: '#c1cf94'},{xInd: 10, yInd: 10, color: '#c1cf94'},{xInd: 10, yInd: 11, color: '#c1cf94'},{xInd: 10, yInd: 12, color: '#c1cf94'},{xInd: 10, yInd: 13, color: '#c1cf94'},{xInd: 10, yInd: 14, color: '#c1cf94'},{xInd: 10, yInd: 15, color: '#c1cf94'},{xInd: 10, yInd: 16, color: '#c1cf94'},{xInd: 10, yInd: 17, color: '#c1cf94'},{xInd: 10, yInd: 18, color: '#c1cf94'},{xInd: 10, yInd: 19, color: '#c1cf94'},{xInd: 10, yInd: 20, color: '#c1cf94'},{xInd: 10, yInd: 21, color: '#487ba6'},{xInd: 10, yInd: 22, color: '#487ba6'},{xInd: 11, yInd: 7, color: '#fed070'},{xInd: 11, yInd: 8, color: '#fed070'},{xInd: 11, yInd: 21, color: '#487ba6'},{xInd: 11, yInd: 22, color: '#487ba6'},{xInd: 12, yInd: 5, color: '#fed070'},{xInd: 12, yInd: 6, color: '#fed070'},{xInd: 12, yInd: 23, color: '#487ba6'},{xInd: 12, yInd: 24, color: '#487ba6'},{xInd: 13, yInd: 5, color: '#fed070'},{xInd: 13, yInd: 6, color: '#fed070'},{xInd: 13, yInd: 23, color: '#487ba6'},{xInd: 13, yInd: 24, color: '#487ba6'},{xInd: 14, yInd: 3, color: '#fed070'},{xInd: 14, yInd: 4, color: '#fed070'},{xInd: 14, yInd: 25, color: '#487ba6'},{xInd: 14, yInd: 26, color: '#487ba6'},{xInd: 15, yInd: 3, color: '#fed070'},{xInd: 15, yInd: 4, color: '#fed070'},{xInd: 15, yInd: 25, color: '#487ba6'},{xInd: 15, yInd: 26, color: '#487ba6'},{xInd: 16, yInd: 1, color: '#fed070'},{xInd: 16, yInd: 2, color: '#fed070'},{xInd: 16, yInd: 27, color: '#487ba6'},{xInd: 16, yInd: 28, color: '#487ba6'},{xInd: 17, yInd: 1, color: '#fed070'},{xInd: 17, yInd: 2, color: '#fed070'},{xInd: 17, yInd: 27, color: '#487ba6'},{xInd: 17, yInd: 28, color: '#487ba6'},],
-                8: [{xInd: 5, yInd: 2, color: '#9f6b53'},{xInd: 5, yInd: 3, color: '#9f6b53'},{xInd: 5, yInd: 4, color: '#9f6b53'},{xInd: 5, yInd: 5, color: '#9f6b53'},{xInd: 5, yInd: 6, color: '#9f6b53'},{xInd: 5, yInd: 7, color: '#9f6b53'},{xInd: 5, yInd: 8, color: '#9f6b53'},{xInd: 5, yInd: 9, color: '#9f6b53'},{xInd: 5, yInd: 10, color: '#9f6b53'},{xInd: 5, yInd: 11, color: '#e38c95'},{xInd: 5, yInd: 12, color: '#e38c95'},{xInd: 5, yInd: 13, color: '#e38c95'},{xInd: 5, yInd: 14, color: '#e38c95'},{xInd: 5, yInd: 15, color: '#e38c95'},{xInd: 5, yInd: 16, color: '#e38c95'},{xInd: 5, yInd: 17, color: '#e38c95'},{xInd: 5, yInd: 18, color: '#e38c95'},{xInd: 5, yInd: 19, color: '#e38c95'},{xInd: 6, yInd: 2, color: '#9f6b53'},{xInd: 6, yInd: 3, color: '#9f6b53'},{xInd: 6, yInd: 4, color: '#9f6b53'},{xInd: 6, yInd: 5, color: '#9f6b53'},{xInd: 6, yInd: 6, color: '#9f6b53'},{xInd: 6, yInd: 7, color: '#9f6b53'},{xInd: 6, yInd: 8, color: '#9f6b53'},{xInd: 6, yInd: 9, color: '#9f6b53'},{xInd: 6, yInd: 10, color: '#9f6b53'},{xInd: 6, yInd: 11, color: '#e38c95'},{xInd: 6, yInd: 12, color: '#e38c95'},{xInd: 6, yInd: 13, color: '#e38c95'},{xInd: 6, yInd: 14, color: '#e38c95'},{xInd: 6, yInd: 15, color: '#e38c95'},{xInd: 6, yInd: 16, color: '#e38c95'},{xInd: 6, yInd: 17, color: '#e38c95'},{xInd: 6, yInd: 18, color: '#e38c95'},{xInd: 6, yInd: 19, color: '#e38c95'},{xInd: 7, yInd: 2, color: '#9f6b53'},{xInd: 7, yInd: 3, color: '#9f6b53'},{xInd: 7, yInd: 4, color: '#9f6b53'},{xInd: 7, yInd: 5, color: '#9f6b53'},{xInd: 7, yInd: 6, color: '#9f6b53'},{xInd: 7, yInd: 7, color: '#9f6b53'},{xInd: 7, yInd: 8, color: '#9f6b53'},{xInd: 7, yInd: 9, color: '#9f6b53'},{xInd: 7, yInd: 10, color: '#9f6b53'},{xInd: 7, yInd: 11, color: '#e38c95'},{xInd: 7, yInd: 12, color: '#e38c95'},{xInd: 7, yInd: 13, color: '#e38c95'},{xInd: 7, yInd: 14, color: '#e38c95'},{xInd: 7, yInd: 15, color: '#e38c95'},{xInd: 7, yInd: 16, color: '#e38c95'},{xInd: 7, yInd: 17, color: '#e38c95'},{xInd: 7, yInd: 18, color: '#e38c95'},{xInd: 7, yInd: 19, color: '#e38c95'},{xInd: 8, yInd: 2, color: '#9f6b53'},{xInd: 8, yInd: 3, color: '#9f6b53'},{xInd: 8, yInd: 4, color: '#9f6b53'},{xInd: 8, yInd: 5, color: '#9f6b53'},{xInd: 8, yInd: 6, color: '#9f6b53'},{xInd: 8, yInd: 7, color: '#9f6b53'},{xInd: 8, yInd: 8, color: '#9f6b53'},{xInd: 8, yInd: 9, color: '#9f6b53'},{xInd: 8, yInd: 10, color: '#9f6b53'},{xInd: 8, yInd: 11, color: '#e38c95'},{xInd: 8, yInd: 12, color: '#e38c95'},{xInd: 8, yInd: 13, color: '#e38c95'},{xInd: 8, yInd: 14, color: '#e38c95'},{xInd: 8, yInd: 15, color: '#e38c95'},{xInd: 8, yInd: 16, color: '#e38c95'},{xInd: 8, yInd: 17, color: '#e38c95'},{xInd: 8, yInd: 18, color: '#e38c95'},{xInd: 8, yInd: 19, color: '#e38c95'},{xInd: 9, yInd: 2, color: '#9f6b53'},{xInd: 9, yInd: 3, color: '#9f6b53'},{xInd: 9, yInd: 4, color: '#9f6b53'},{xInd: 9, yInd: 5, color: '#9f6b53'},{xInd: 9, yInd: 6, color: '#9f6b53'},{xInd: 9, yInd: 7, color: '#9f6b53'},{xInd: 9, yInd: 8, color: '#9f6b53'},{xInd: 9, yInd: 9, color: '#9f6b53'},{xInd: 9, yInd: 10, color: '#9f6b53'},{xInd: 9, yInd: 11, color: '#e38c95'},{xInd: 9, yInd: 12, color: '#e38c95'},{xInd: 9, yInd: 13, color: '#e38c95'},{xInd: 9, yInd: 14, color: '#e38c95'},{xInd: 9, yInd: 15, color: '#e38c95'},{xInd: 9, yInd: 16, color: '#e38c95'},{xInd: 9, yInd: 17, color: '#e38c95'},{xInd: 9, yInd: 18, color: '#e38c95'},{xInd: 9, yInd: 19, color: '#e38c95'},{xInd: 10, yInd: 2, color: '#9f6b53'},{xInd: 10, yInd: 3, color: '#9f6b53'},{xInd: 10, yInd: 4, color: '#9f6b53'},{xInd: 10, yInd: 5, color: '#9f6b53'},{xInd: 10, yInd: 6, color: '#9f6b53'},{xInd: 10, yInd: 7, color: '#9f6b53'},{xInd: 10, yInd: 8, color: '#9f6b53'},{xInd: 10, yInd: 9, color: '#9f6b53'},{xInd: 10, yInd: 10, color: '#9f6b53'},{xInd: 10, yInd: 11, color: '#e38c95'},{xInd: 10, yInd: 12, color: '#e38c95'},{xInd: 10, yInd: 13, color: '#e38c95'},{xInd: 10, yInd: 14, color: '#e38c95'},{xInd: 10, yInd: 15, color: '#e38c95'},{xInd: 10, yInd: 16, color: '#e38c95'},{xInd: 10, yInd: 17, color: '#e38c95'},{xInd: 10, yInd: 18, color: '#e38c95'},{xInd: 10, yInd: 19, color: '#e38c95'},{xInd: 11, yInd: 2, color: '#9f6b53'},{xInd: 11, yInd: 3, color: '#9f6b53'},{xInd: 11, yInd: 4, color: '#9f6b53'},{xInd: 11, yInd: 5, color: '#9f6b53'},{xInd: 11, yInd: 6, color: '#9f6b53'},{xInd: 11, yInd: 7, color: '#9f6b53'},{xInd: 11, yInd: 8, color: '#9f6b53'},{xInd: 11, yInd: 9, color: '#9f6b53'},{xInd: 11, yInd: 10, color: '#9f6b53'},{xInd: 11, yInd: 11, color: '#e38c95'},{xInd: 11, yInd: 12, color: '#e38c95'},{xInd: 11, yInd: 13, color: '#e38c95'},{xInd: 11, yInd: 14, color: '#e38c95'},{xInd: 11, yInd: 15, color: '#e38c95'},{xInd: 11, yInd: 16, color: '#e38c95'},{xInd: 11, yInd: 17, color: '#e38c95'},{xInd: 11, yInd: 18, color: '#e38c95'},{xInd: 11, yInd: 19, color: '#e38c95'},{xInd: 12, yInd: 2, color: '#9f6b53'},{xInd: 12, yInd: 3, color: '#9f6b53'},{xInd: 12, yInd: 4, color: '#9f6b53'},{xInd: 12, yInd: 5, color: '#9f6b53'},{xInd: 12, yInd: 6, color: '#9f6b53'},{xInd: 12, yInd: 7, color: '#9f6b53'},{xInd: 12, yInd: 8, color: '#9f6b53'},{xInd: 12, yInd: 9, color: '#9f6b53'},{xInd: 12, yInd: 10, color: '#9f6b53'},{xInd: 12, yInd: 11, color: '#e38c95'},{xInd: 12, yInd: 12, color: '#e38c95'},{xInd: 12, yInd: 13, color: '#e38c95'},{xInd: 12, yInd: 14, color: '#e38c95'},{xInd: 12, yInd: 15, color: '#e38c95'},{xInd: 12, yInd: 16, color: '#e38c95'},{xInd: 12, yInd: 17, color: '#e38c95'},{xInd: 12, yInd: 18, color: '#e38c95'},{xInd: 12, yInd: 19, color: '#e38c95'},{xInd: 13, yInd: 2, color: '#9f6b53'},{xInd: 13, yInd: 3, color: '#9f6b53'},{xInd: 13, yInd: 4, color: '#9f6b53'},{xInd: 13, yInd: 5, color: '#9f6b53'},{xInd: 13, yInd: 6, color: '#9f6b53'},{xInd: 13, yInd: 7, color: '#9f6b53'},{xInd: 13, yInd: 8, color: '#9f6b53'},{xInd: 13, yInd: 9, color: '#9f6b53'},{xInd: 13, yInd: 10, color: '#9f6b53'},{xInd: 13, yInd: 11, color: '#e38c95'},{xInd: 13, yInd: 12, color: '#e38c95'},{xInd: 13, yInd: 13, color: '#e38c95'},{xInd: 13, yInd: 14, color: '#e38c95'},{xInd: 13, yInd: 15, color: '#e38c95'},{xInd: 13, yInd: 16, color: '#e38c95'},{xInd: 13, yInd: 17, color: '#e38c95'},{xInd: 13, yInd: 18, color: '#e38c95'},{xInd: 13, yInd: 19, color: '#e38c95'},],
-9: [{xInd: 1, yInd: 6, color: '#e68b86'},{xInd: 1, yInd: 7, color: '#e68b86'},{xInd: 1, yInd: 14, color: '#e68b86'},{xInd: 1, yInd: 15, color: '#e68b86'},{xInd: 1, yInd: 22, color: '#e68b86'},{xInd: 1, yInd: 23, color: '#e68b86'},{xInd: 2, yInd: 6, color: '#e68b86'},{xInd: 2, yInd: 7, color: '#e68b86'},{xInd: 2, yInd: 14, color: '#e68b86'},{xInd: 2, yInd: 15, color: '#e68b86'},{xInd: 2, yInd: 22, color: '#e68b86'},{xInd: 2, yInd: 23, color: '#e68b86'},{xInd: 3, yInd: 6, color: '#e68b86'},{xInd: 3, yInd: 7, color: '#e68b86'},{xInd: 3, yInd: 14, color: '#e68b86'},{xInd: 3, yInd: 15, color: '#e68b86'},{xInd: 3, yInd: 22, color: '#e68b86'},{xInd: 3, yInd: 23, color: '#e68b86'},{xInd: 4, yInd: 6, color: '#e68b86'},{xInd: 4, yInd: 7, color: '#e68b86'},{xInd: 4, yInd: 14, color: '#e68b86'},{xInd: 4, yInd: 15, color: '#e68b86'},{xInd: 4, yInd: 22, color: '#e68b86'},{xInd: 4, yInd: 23, color: '#e68b86'},{xInd: 5, yInd: 6, color: '#e68b86'},{xInd: 5, yInd: 7, color: '#e68b86'},{xInd: 5, yInd: 14, color: '#e68b86'},{xInd: 5, yInd: 15, color: '#e68b86'},{xInd: 5, yInd: 22, color: '#e68b86'},{xInd: 5, yInd: 23, color: '#e68b86'},{xInd: 6, yInd: 6, color: '#e68b86'},{xInd: 6, yInd: 7, color: '#e68b86'},{xInd: 6, yInd: 14, color: '#e68b86'},{xInd: 6, yInd: 15, color: '#e68b86'},{xInd: 6, yInd: 22, color: '#e68b86'},{xInd: 6, yInd: 23, color: '#e68b86'},{xInd: 7, yInd: 6, color: '#e68b86'},{xInd: 7, yInd: 7, color: '#e68b86'},{xInd: 7, yInd: 14, color: '#e68b86'},{xInd: 7, yInd: 15, color: '#e68b86'},{xInd: 7, yInd: 22, color: '#e68b86'},{xInd: 7, yInd: 23, color: '#e68b86'},{xInd: 8, yInd: 6, color: '#e68b86'},{xInd: 8, yInd: 7, color: '#e68b86'},{xInd: 8, yInd: 14, color: '#e68b86'},{xInd: 8, yInd: 15, color: '#e68b86'},{xInd: 8, yInd: 22, color: '#e68b86'},{xInd: 8, yInd: 23, color: '#e68b86'},{xInd: 9, yInd: 2, color: '#c9b4bb'},{xInd: 9, yInd: 3, color: '#c9b4bb'},{xInd: 9, yInd: 4, color: '#c9b4bb'},{xInd: 9, yInd: 5, color: '#c9b4bb'},{xInd: 9, yInd: 6, color: '#c9b4bb'},{xInd: 9, yInd: 7, color: '#c9b4bb'},{xInd: 9, yInd: 8, color: '#c9b4bb'},{xInd: 9, yInd: 9, color: '#c9b4bb'},{xInd: 9, yInd: 10, color: '#c9b4bb'},{xInd: 9, yInd: 11, color: '#c9b4bb'},{xInd: 9, yInd: 12, color: '#c9b4bb'},{xInd: 9, yInd: 13, color: '#c9b4bb'},{xInd: 9, yInd: 14, color: '#c9b4bb'},{xInd: 9, yInd: 15, color: '#c9b4bb'},{xInd: 9, yInd: 16, color: '#c9b4bb'},{xInd: 9, yInd: 17, color: '#c9b4bb'},{xInd: 9, yInd: 18, color: '#c9b4bb'},{xInd: 9, yInd: 19, color: '#c9b4bb'},{xInd: 9, yInd: 20, color: '#c9b4bb'},{xInd: 9, yInd: 21, color: '#c9b4bb'},{xInd: 9, yInd: 22, color: '#c9b4bb'},{xInd: 9, yInd: 23, color: '#c9b4bb'},{xInd: 9, yInd: 24, color: '#c9b4bb'},{xInd: 9, yInd: 25, color: '#c9b4bb'},{xInd: 9, yInd: 26, color: '#c9b4bb'},{xInd: 9, yInd: 27, color: '#c9b4bb'},{xInd: 10, yInd: 2, color: '#5d97af'},{xInd: 10, yInd: 3, color: '#5d97af'},{xInd: 10, yInd: 10, color: '#5d97af'},{xInd: 10, yInd: 11, color: '#5d97af'},{xInd: 10, yInd: 18, color: '#5d97af'},{xInd: 10, yInd: 19, color: '#5d97af'},{xInd: 10, yInd: 26, color: '#5d97af'},{xInd: 10, yInd: 27, color: '#5d97af'},{xInd: 11, yInd: 2, color: '#5d97af'},{xInd: 11, yInd: 3, color: '#5d97af'},{xInd: 11, yInd: 10, color: '#5d97af'},{xInd: 11, yInd: 11, color: '#5d97af'},{xInd: 11, yInd: 18, color: '#5d97af'},{xInd: 11, yInd: 19, color: '#5d97af'},{xInd: 11, yInd: 26, color: '#5d97af'},{xInd: 11, yInd: 27, color: '#5d97af'},{xInd: 12, yInd: 2, color: '#5d97af'},{xInd: 12, yInd: 3, color: '#5d97af'},{xInd: 12, yInd: 10, color: '#5d97af'},{xInd: 12, yInd: 11, color: '#5d97af'},{xInd: 12, yInd: 18, color: '#5d97af'},{xInd: 12, yInd: 19, color: '#5d97af'},{xInd: 12, yInd: 26, color: '#5d97af'},{xInd: 12, yInd: 27, color: '#5d97af'},{xInd: 13, yInd: 2, color: '#5d97af'},{xInd: 13, yInd: 3, color: '#5d97af'},{xInd: 13, yInd: 10, color: '#5d97af'},{xInd: 13, yInd: 11, color: '#5d97af'},{xInd: 13, yInd: 18, color: '#5d97af'},{xInd: 13, yInd: 19, color: '#5d97af'},{xInd: 13, yInd: 26, color: '#5d97af'},{xInd: 13, yInd: 27, color: '#5d97af'},{xInd: 14, yInd: 2, color: '#5d97af'},{xInd: 14, yInd: 3, color: '#5d97af'},{xInd: 14, yInd: 10, color: '#5d97af'},{xInd: 14, yInd: 11, color: '#5d97af'},{xInd: 14, yInd: 18, color: '#5d97af'},{xInd: 14, yInd: 19, color: '#5d97af'},{xInd: 14, yInd: 26, color: '#5d97af'},{xInd: 14, yInd: 27, color: '#5d97af'},{xInd: 15, yInd: 2, color: '#5d97af'},{xInd: 15, yInd: 3, color: '#5d97af'},{xInd: 15, yInd: 10, color: '#5d97af'},{xInd: 15, yInd: 11, color: '#5d97af'},{xInd: 15, yInd: 18, color: '#5d97af'},{xInd: 15, yInd: 19, color: '#5d97af'},{xInd: 15, yInd: 26, color: '#5d97af'},{xInd: 15, yInd: 27, color: '#5d97af'},{xInd: 16, yInd: 2, color: '#5d97af'},{xInd: 16, yInd: 3, color: '#5d97af'},{xInd: 16, yInd: 10, color: '#5d97af'},{xInd: 16, yInd: 11, color: '#5d97af'},{xInd: 16, yInd: 18, color: '#5d97af'},{xInd: 16, yInd: 19, color: '#5d97af'},{xInd: 16, yInd: 26, color: '#5d97af'},{xInd: 16, yInd: 27, color: '#5d97af'},{xInd: 17, yInd: 2, color: '#5d97af'},{xInd: 17, yInd: 3, color: '#5d97af'},{xInd: 17, yInd: 10, color: '#5d97af'},{xInd: 17, yInd: 11, color: '#5d97af'},{xInd: 17, yInd: 18, color: '#5d97af'},{xInd: 17, yInd: 19, color: '#5d97af'},{xInd: 17, yInd: 26, color: '#5d97af'},{xInd: 17, yInd: 27, color: '#5d97af'},],
-10: [{xInd: 0, yInd: 11, color: '#94f2f4'},{xInd: 0, yInd: 12, color: '#f5deb2'},{xInd: 0, yInd: 13, color: '#f5deb2'},{xInd: 0, yInd: 14, color: '#f5deb2'},{xInd: 0, yInd: 15, color: '#f5deb2'},{xInd: 0, yInd: 16, color: '#f5deb2'},{xInd: 0, yInd: 17, color: '#f5deb2'},{xInd: 0, yInd: 18, color: '#f5deb2'},{xInd: 0, yInd: 19, color: '#f5deb2'},{xInd: 0, yInd: 20, color: '#f5deb2'},{xInd: 0, yInd: 21, color: '#f5deb2'},{xInd: 0, yInd: 22, color: '#f5deb2'},{xInd: 0, yInd: 23, color: '#f5deb2'},{xInd: 0, yInd: 24, color: '#f5deb2'},{xInd: 0, yInd: 25, color: '#f5deb2'},{xInd: 0, yInd: 26, color: '#f5deb2'},{xInd: 1, yInd: 10, color: '#94f2f4'},{xInd: 1, yInd: 11, color: '#94f2f4'},{xInd: 1, yInd: 12, color: '#f5deb2'},{xInd: 1, yInd: 13, color: '#f5deb2'},{xInd: 1, yInd: 14, color: '#f5deb2'},{xInd: 1, yInd: 15, color: '#f5deb2'},{xInd: 1, yInd: 16, color: '#f5deb2'},{xInd: 1, yInd: 17, color: '#f5deb2'},{xInd: 1, yInd: 18, color: '#f5deb2'},{xInd: 1, yInd: 19, color: '#f5deb2'},{xInd: 1, yInd: 20, color: '#f5deb2'},{xInd: 1, yInd: 21, color: '#f5deb2'},{xInd: 1, yInd: 22, color: '#f5deb2'},{xInd: 1, yInd: 23, color: '#f5deb2'},{xInd: 1, yInd: 24, color: '#f5deb2'},{xInd: 1, yInd: 25, color: '#f5deb2'},{xInd: 1, yInd: 26, color: '#f5deb2'},{xInd: 2, yInd: 9, color: '#94f2f4'},{xInd: 2, yInd: 10, color: '#94f2f4'},{xInd: 2, yInd: 11, color: '#94f2f4'},{xInd: 2, yInd: 12, color: '#f5deb2'},{xInd: 2, yInd: 13, color: '#f5deb2'},{xInd: 2, yInd: 14, color: '#f5deb2'},{xInd: 2, yInd: 15, color: '#f5deb2'},{xInd: 2, yInd: 16, color: '#f5deb2'},{xInd: 2, yInd: 17, color: '#f5deb2'},{xInd: 2, yInd: 18, color: '#f5deb2'},{xInd: 2, yInd: 19, color: '#f5deb2'},{xInd: 2, yInd: 20, color: '#f5deb2'},{xInd: 2, yInd: 21, color: '#f5deb2'},{xInd: 2, yInd: 22, color: '#f5deb2'},{xInd: 2, yInd: 23, color: '#f5deb2'},{xInd: 2, yInd: 24, color: '#f5deb2'},{xInd: 2, yInd: 25, color: '#f5deb2'},{xInd: 2, yInd: 26, color: '#f5deb2'},{xInd: 3, yInd: 8, color: '#94f2f4'},{xInd: 3, yInd: 9, color: '#94f2f4'},{xInd: 3, yInd: 10, color: '#94f2f4'},{xInd: 3, yInd: 11, color: '#94f2f4'},{xInd: 3, yInd: 12, color: '#f5deb2'},{xInd: 3, yInd: 13, color: '#f5deb2'},{xInd: 3, yInd: 14, color: '#f5deb2'},{xInd: 3, yInd: 15, color: '#f5deb2'},{xInd: 3, yInd: 16, color: '#f5deb2'},{xInd: 3, yInd: 17, color: '#f5deb2'},{xInd: 3, yInd: 18, color: '#f5deb2'},{xInd: 3, yInd: 19, color: '#f5deb2'},{xInd: 3, yInd: 20, color: '#f5deb2'},{xInd: 3, yInd: 21, color: '#f5deb2'},{xInd: 3, yInd: 22, color: '#f5deb2'},{xInd: 3, yInd: 23, color: '#f5deb2'},{xInd: 3, yInd: 24, color: '#f5deb2'},{xInd: 3, yInd: 25, color: '#f5deb2'},{xInd: 3, yInd: 26, color: '#f5deb2'},{xInd: 4, yInd: 7, color: '#94f2f4'},{xInd: 4, yInd: 8, color: '#94f2f4'},{xInd: 4, yInd: 9, color: '#94f2f4'},{xInd: 4, yInd: 10, color: '#94f2f4'},{xInd: 4, yInd: 11, color: '#94f2f4'},{xInd: 5, yInd: 6, color: '#88adb3'},{xInd: 5, yInd: 7, color: '#94f2f4'},{xInd: 5, yInd: 8, color: '#94f2f4'},{xInd: 5, yInd: 9, color: '#94f2f4'},{xInd: 5, yInd: 10, color: '#94f2f4'},{xInd: 5, yInd: 11, color: '#94f2f4'},{xInd: 6, yInd: 5, color: '#88adb3'},{xInd: 6, yInd: 6, color: '#88adb3'},{xInd: 6, yInd: 7, color: '#94f2f4'},{xInd: 6, yInd: 8, color: '#94f2f4'},{xInd: 6, yInd: 9, color: '#94f2f4'},{xInd: 6, yInd: 10, color: '#94f2f4'},{xInd: 6, yInd: 11, color: '#94f2f4'},{xInd: 7, yInd: 4, color: '#88adb3'},{xInd: 7, yInd: 5, color: '#88adb3'},{xInd: 7, yInd: 6, color: '#88adb3'},{xInd: 7, yInd: 7, color: '#94f2f4'},{xInd: 7, yInd: 8, color: '#94f2f4'},{xInd: 7, yInd: 9, color: '#94f2f4'},{xInd: 7, yInd: 10, color: '#94f2f4'},{xInd: 7, yInd: 11, color: '#94f2f4'},{xInd: 8, yInd: 3, color: '#88adb3'},{xInd: 8, yInd: 4, color: '#88adb3'},{xInd: 8, yInd: 5, color: '#88adb3'},{xInd: 8, yInd: 6, color: '#88adb3'},{xInd: 8, yInd: 7, color: '#94f2f4'},{xInd: 8, yInd: 8, color: '#94f2f4'},{xInd: 8, yInd: 9, color: '#94f2f4'},{xInd: 8, yInd: 10, color: '#94f2f4'},{xInd: 8, yInd: 11, color: '#94f2f4'},{xInd: 9, yInd: 2, color: '#88adb3'},{xInd: 9, yInd: 3, color: '#88adb3'},{xInd: 9, yInd: 4, color: '#88adb3'},{xInd: 9, yInd: 5, color: '#88adb3'},{xInd: 9, yInd: 6, color: '#88adb3'},{xInd: 9, yInd: 7, color: '#94f2f4'},{xInd: 9, yInd: 8, color: '#94f2f4'},{xInd: 9, yInd: 9, color: '#94f2f4'},{xInd: 9, yInd: 10, color: '#94f2f4'},{xInd: 9, yInd: 11, color: '#94f2f4'},{xInd: 10, yInd: 3, color: '#88adb3'},{xInd: 10, yInd: 4, color: '#88adb3'},{xInd: 10, yInd: 5, color: '#88adb3'},{xInd: 10, yInd: 6, color: '#88adb3'},{xInd: 10, yInd: 7, color: '#94f2f4'},{xInd: 10, yInd: 8, color: '#94f2f4'},{xInd: 10, yInd: 9, color: '#94f2f4'},{xInd: 10, yInd: 10, color: '#94f2f4'},{xInd: 10, yInd: 11, color: '#94f2f4'},{xInd: 11, yInd: 4, color: '#88adb3'},{xInd: 11, yInd: 5, color: '#88adb3'},{xInd: 11, yInd: 6, color: '#88adb3'},{xInd: 11, yInd: 7, color: '#94f2f4'},{xInd: 11, yInd: 8, color: '#94f2f4'},{xInd: 11, yInd: 9, color: '#94f2f4'},{xInd: 11, yInd: 10, color: '#94f2f4'},{xInd: 11, yInd: 11, color: '#94f2f4'},{xInd: 12, yInd: 5, color: '#88adb3'},{xInd: 12, yInd: 6, color: '#88adb3'},{xInd: 12, yInd: 7, color: '#94f2f4'},{xInd: 12, yInd: 8, color: '#94f2f4'},{xInd: 12, yInd: 9, color: '#94f2f4'},{xInd: 12, yInd: 10, color: '#94f2f4'},{xInd: 12, yInd: 11, color: '#94f2f4'},{xInd: 13, yInd: 6, color: '#88adb3'},{xInd: 13, yInd: 7, color: '#94f2f4'},{xInd: 13, yInd: 8, color: '#94f2f4'},{xInd: 13, yInd: 9, color: '#94f2f4'},{xInd: 13, yInd: 10, color: '#94f2f4'},{xInd: 13, yInd: 11, color: '#94f2f4'},{xInd: 14, yInd: 7, color: '#94f2f4'},{xInd: 14, yInd: 8, color: '#94f2f4'},{xInd: 14, yInd: 9, color: '#94f2f4'},{xInd: 14, yInd: 10, color: '#94f2f4'},{xInd: 14, yInd: 11, color: '#94f2f4'},{xInd: 15, yInd: 8, color: '#94f2f4'},{xInd: 15, yInd: 9, color: '#94f2f4'},{xInd: 15, yInd: 10, color: '#94f2f4'},{xInd: 15, yInd: 11, color: '#94f2f4'},{xInd: 15, yInd: 12, color: '#f5deb2'},{xInd: 15, yInd: 13, color: '#f5deb2'},{xInd: 15, yInd: 14, color: '#f5deb2'},{xInd: 15, yInd: 15, color: '#f5deb2'},{xInd: 15, yInd: 16, color: '#f5deb2'},{xInd: 15, yInd: 17, color: '#f5deb2'},{xInd: 15, yInd: 18, color: '#f5deb2'},{xInd: 15, yInd: 19, color: '#f5deb2'},{xInd: 15, yInd: 20, color: '#f5deb2'},{xInd: 15, yInd: 21, color: '#f5deb2'},{xInd: 15, yInd: 22, color: '#f5deb2'},{xInd: 15, yInd: 23, color: '#f5deb2'},{xInd: 15, yInd: 24, color: '#f5deb2'},{xInd: 15, yInd: 25, color: '#f5deb2'},{xInd: 15, yInd: 26, color: '#f5deb2'},{xInd: 16, yInd: 9, color: '#94f2f4'},{xInd: 16, yInd: 10, color: '#94f2f4'},{xInd: 16, yInd: 11, color: '#94f2f4'},{xInd: 16, yInd: 12, color: '#f5deb2'},{xInd: 16, yInd: 13, color: '#f5deb2'},{xInd: 16, yInd: 14, color: '#f5deb2'},{xInd: 16, yInd: 15, color: '#f5deb2'},{xInd: 16, yInd: 16, color: '#f5deb2'},{xInd: 16, yInd: 17, color: '#f5deb2'},{xInd: 16, yInd: 18, color: '#f5deb2'},{xInd: 16, yInd: 19, color: '#f5deb2'},{xInd: 16, yInd: 20, color: '#f5deb2'},{xInd: 16, yInd: 21, color: '#f5deb2'},{xInd: 16, yInd: 22, color: '#f5deb2'},{xInd: 16, yInd: 23, color: '#f5deb2'},{xInd: 16, yInd: 24, color: '#f5deb2'},{xInd: 16, yInd: 25, color: '#f5deb2'},{xInd: 16, yInd: 26, color: '#f5deb2'},{xInd: 17, yInd: 10, color: '#94f2f4'},{xInd: 17, yInd: 11, color: '#94f2f4'},{xInd: 17, yInd: 12, color: '#f5deb2'},{xInd: 17, yInd: 13, color: '#f5deb2'},{xInd: 17, yInd: 14, color: '#f5deb2'},{xInd: 17, yInd: 15, color: '#f5deb2'},{xInd: 17, yInd: 16, color: '#f5deb2'},{xInd: 17, yInd: 17, color: '#f5deb2'},{xInd: 17, yInd: 18, color: '#f5deb2'},{xInd: 17, yInd: 19, color: '#f5deb2'},{xInd: 17, yInd: 20, color: '#f5deb2'},{xInd: 17, yInd: 21, color: '#f5deb2'},{xInd: 17, yInd: 22, color: '#f5deb2'},{xInd: 17, yInd: 23, color: '#f5deb2'},{xInd: 17, yInd: 24, color: '#f5deb2'},{xInd: 17, yInd: 25, color: '#f5deb2'},{xInd: 17, yInd: 26, color: '#f5deb2'},{xInd: 18, yInd: 11, color: '#94f2f4'},{xInd: 18, yInd: 12, color: '#f5deb2'},{xInd: 18, yInd: 13, color: '#f5deb2'},{xInd: 18, yInd: 14, color: '#f5deb2'},{xInd: 18, yInd: 15, color: '#f5deb2'},{xInd: 18, yInd: 16, color: '#f5deb2'},{xInd: 18, yInd: 17, color: '#f5deb2'},{xInd: 18, yInd: 18, color: '#f5deb2'},{xInd: 18, yInd: 19, color: '#f5deb2'},{xInd: 18, yInd: 20, color: '#f5deb2'},{xInd: 18, yInd: 21, color: '#f5deb2'},{xInd: 18, yInd: 22, color: '#f5deb2'},{xInd: 18, yInd: 23, color: '#f5deb2'},{xInd: 18, yInd: 24, color: '#f5deb2'},{xInd: 18, yInd: 25, color: '#f5deb2'},{xInd: 18, yInd: 26, color: '#f5deb2'},],
-11: [{xInd: 0, yInd: 10, color: '#f2a557'},{xInd: 0, yInd: 11, color: '#f2a557'},{xInd: 0, yInd: 12, color: '#f2a557'},{xInd: 0, yInd: 13, color: '#f2a557'},{xInd: 0, yInd: 14, color: '#f2a557'},{xInd: 0, yInd: 15, color: '#f2a557'},{xInd: 0, yInd: 16, color: '#f2a557'},{xInd: 0, yInd: 17, color: '#f2a557'},{xInd: 0, yInd: 18, color: '#f2a557'},{xInd: 0, yInd: 19, color: '#f2a557'},{xInd: 0, yInd: 20, color: '#f2a557'},{xInd: 0, yInd: 21, color: '#f2a557'},{xInd: 0, yInd: 22, color: '#f2a557'},{xInd: 0, yInd: 23, color: '#f2a557'},{xInd: 0, yInd: 24, color: '#f2a557'},{xInd: 0, yInd: 25, color: '#f2a557'},{xInd: 0, yInd: 26, color: '#f2a557'},{xInd: 0, yInd: 27, color: '#f2a557'},{xInd: 0, yInd: 28, color: '#f2a557'},{xInd: 1, yInd: 12, color: '#e7b74b'},{xInd: 1, yInd: 13, color: '#e7b74b'},{xInd: 1, yInd: 14, color: '#e7b74b'},{xInd: 1, yInd: 15, color: '#e7b74b'},{xInd: 1, yInd: 16, color: '#e7b74b'},{xInd: 1, yInd: 17, color: '#e7b74b'},{xInd: 1, yInd: 18, color: '#e7b74b'},{xInd: 1, yInd: 19, color: '#e7b74b'},{xInd: 1, yInd: 20, color: '#e7b74b'},{xInd: 1, yInd: 21, color: '#e7b74b'},{xInd: 1, yInd: 22, color: '#e7b74b'},{xInd: 1, yInd: 23, color: '#e7b74b'},{xInd: 1, yInd: 24, color: '#e7b74b'},{xInd: 1, yInd: 25, color: '#e7b74b'},{xInd: 1, yInd: 26, color: '#e7b74b'},{xInd: 1, yInd: 27, color: '#e7b74b'},{xInd: 1, yInd: 28, color: '#f2a557'},{xInd: 2, yInd: 10, color: '#f2a557'},{xInd: 2, yInd: 11, color: '#f2a557'},{xInd: 2, yInd: 12, color: '#f2a557'},{xInd: 2, yInd: 13, color: '#f2a557'},{xInd: 2, yInd: 14, color: '#f2a557'},{xInd: 2, yInd: 15, color: '#f2a557'},{xInd: 2, yInd: 16, color: '#f2a557'},{xInd: 2, yInd: 17, color: '#f2a557'},{xInd: 2, yInd: 18, color: '#f2a557'},{xInd: 2, yInd: 19, color: '#f2a557'},{xInd: 2, yInd: 20, color: '#f2a557'},{xInd: 2, yInd: 21, color: '#f2a557'},{xInd: 2, yInd: 22, color: '#f2a557'},{xInd: 2, yInd: 23, color: '#f2a557'},{xInd: 2, yInd: 24, color: '#f2a557'},{xInd: 2, yInd: 25, color: '#f2a557'},{xInd: 2, yInd: 26, color: '#f2a557'},{xInd: 2, yInd: 27, color: '#f2a557'},{xInd: 2, yInd: 28, color: '#f2a557'},{xInd: 3, yInd: 12, color: '#e7b74b'},{xInd: 3, yInd: 13, color: '#e7b74b'},{xInd: 3, yInd: 14, color: '#e7b74b'},{xInd: 3, yInd: 15, color: '#e7b74b'},{xInd: 3, yInd: 16, color: '#e7b74b'},{xInd: 3, yInd: 17, color: '#e7b74b'},{xInd: 3, yInd: 18, color: '#e7b74b'},{xInd: 3, yInd: 19, color: '#e7b74b'},{xInd: 3, yInd: 20, color: '#e7b74b'},{xInd: 3, yInd: 21, color: '#e7b74b'},{xInd: 3, yInd: 22, color: '#e7b74b'},{xInd: 3, yInd: 23, color: '#e7b74b'},{xInd: 3, yInd: 24, color: '#e7b74b'},{xInd: 3, yInd: 25, color: '#e7b74b'},{xInd: 3, yInd: 26, color: '#e7b74b'},{xInd: 3, yInd: 27, color: '#e7b74b'},{xInd: 3, yInd: 28, color: '#f2a557'},{xInd: 4, yInd: 10, color: '#f2a557'},{xInd: 4, yInd: 11, color: '#f2a557'},{xInd: 4, yInd: 12, color: '#f2a557'},{xInd: 4, yInd: 13, color: '#f2a557'},{xInd: 4, yInd: 14, color: '#f2a557'},{xInd: 4, yInd: 15, color: '#f2a557'},{xInd: 4, yInd: 16, color: '#f2a557'},{xInd: 4, yInd: 17, color: '#f2a557'},{xInd: 4, yInd: 18, color: '#f2a557'},{xInd: 4, yInd: 19, color: '#f2a557'},{xInd: 4, yInd: 20, color: '#f2a557'},{xInd: 4, yInd: 21, color: '#f2a557'},{xInd: 4, yInd: 22, color: '#f2a557'},{xInd: 4, yInd: 23, color: '#f2a557'},{xInd: 4, yInd: 24, color: '#f2a557'},{xInd: 4, yInd: 25, color: '#f2a557'},{xInd: 4, yInd: 26, color: '#f2a557'},{xInd: 4, yInd: 27, color: '#f2a557'},{xInd: 4, yInd: 28, color: '#f2a557'},{xInd: 5, yInd: 12, color: '#e7b74b'},{xInd: 5, yInd: 13, color: '#e7b74b'},{xInd: 5, yInd: 14, color: '#e7b74b'},{xInd: 5, yInd: 15, color: '#e7b74b'},{xInd: 5, yInd: 16, color: '#e7b74b'},{xInd: 5, yInd: 17, color: '#e7b74b'},{xInd: 5, yInd: 18, color: '#e7b74b'},{xInd: 5, yInd: 19, color: '#e7b74b'},{xInd: 5, yInd: 20, color: '#e7b74b'},{xInd: 5, yInd: 21, color: '#e7b74b'},{xInd: 5, yInd: 22, color: '#e7b74b'},{xInd: 5, yInd: 23, color: '#e7b74b'},{xInd: 5, yInd: 24, color: '#e7b74b'},{xInd: 5, yInd: 25, color: '#e7b74b'},{xInd: 5, yInd: 26, color: '#e7b74b'},{xInd: 5, yInd: 27, color: '#e7b74b'},{xInd: 5, yInd: 28, color: '#f2a557'},{xInd: 6, yInd: 10, color: '#f2a557'},{xInd: 6, yInd: 11, color: '#f2a557'},{xInd: 6, yInd: 12, color: '#f2a557'},{xInd: 6, yInd: 13, color: '#f2a557'},{xInd: 6, yInd: 14, color: '#f2a557'},{xInd: 6, yInd: 15, color: '#f2a557'},{xInd: 6, yInd: 16, color: '#f2a557'},{xInd: 6, yInd: 17, color: '#f2a557'},{xInd: 6, yInd: 18, color: '#f2a557'},{xInd: 6, yInd: 19, color: '#f2a557'},{xInd: 6, yInd: 20, color: '#f2a557'},{xInd: 6, yInd: 21, color: '#f2a557'},{xInd: 6, yInd: 22, color: '#f2a557'},{xInd: 6, yInd: 23, color: '#f2a557'},{xInd: 6, yInd: 24, color: '#f2a557'},{xInd: 6, yInd: 25, color: '#f2a557'},{xInd: 6, yInd: 26, color: '#f2a557'},{xInd: 6, yInd: 27, color: '#f2a557'},{xInd: 6, yInd: 28, color: '#f2a557'},{xInd: 7, yInd: 12, color: '#e7b74b'},{xInd: 7, yInd: 13, color: '#e7b74b'},{xInd: 7, yInd: 14, color: '#e7b74b'},{xInd: 7, yInd: 15, color: '#e7b74b'},{xInd: 7, yInd: 16, color: '#e7b74b'},{xInd: 7, yInd: 17, color: '#e7b74b'},{xInd: 7, yInd: 18, color: '#e7b74b'},{xInd: 7, yInd: 19, color: '#e7b74b'},{xInd: 7, yInd: 20, color: '#e7b74b'},{xInd: 7, yInd: 21, color: '#e7b74b'},{xInd: 7, yInd: 22, color: '#e7b74b'},{xInd: 7, yInd: 23, color: '#e7b74b'},{xInd: 7, yInd: 24, color: '#e7b74b'},{xInd: 7, yInd: 25, color: '#e7b74b'},{xInd: 7, yInd: 26, color: '#e7b74b'},{xInd: 7, yInd: 27, color: '#e7b74b'},{xInd: 7, yInd: 28, color: '#f2a557'},{xInd: 8, yInd: 10, color: '#f2a557'},{xInd: 8, yInd: 11, color: '#f2a557'},{xInd: 8, yInd: 12, color: '#f2a557'},{xInd: 8, yInd: 13, color: '#f2a557'},{xInd: 8, yInd: 14, color: '#f2a557'},{xInd: 8, yInd: 15, color: '#f2a557'},{xInd: 8, yInd: 16, color: '#f2a557'},{xInd: 8, yInd: 17, color: '#f2a557'},{xInd: 8, yInd: 18, color: '#f2a557'},{xInd: 8, yInd: 19, color: '#f2a557'},{xInd: 8, yInd: 20, color: '#f2a557'},{xInd: 8, yInd: 21, color: '#f2a557'},{xInd: 8, yInd: 22, color: '#f2a557'},{xInd: 8, yInd: 23, color: '#f2a557'},{xInd: 8, yInd: 24, color: '#f2a557'},{xInd: 8, yInd: 25, color: '#f2a557'},{xInd: 8, yInd: 26, color: '#f2a557'},{xInd: 8, yInd: 27, color: '#f2a557'},{xInd: 8, yInd: 28, color: '#f2a557'},{xInd: 9, yInd: 12, color: '#e7b74b'},{xInd: 9, yInd: 13, color: '#e7b74b'},{xInd: 9, yInd: 14, color: '#e7b74b'},{xInd: 9, yInd: 15, color: '#e7b74b'},{xInd: 9, yInd: 16, color: '#e7b74b'},{xInd: 9, yInd: 17, color: '#e7b74b'},{xInd: 9, yInd: 18, color: '#e7b74b'},{xInd: 9, yInd: 19, color: '#e7b74b'},{xInd: 9, yInd: 20, color: '#e7b74b'},{xInd: 9, yInd: 21, color: '#e7b74b'},{xInd: 9, yInd: 22, color: '#e7b74b'},{xInd: 9, yInd: 23, color: '#e7b74b'},{xInd: 9, yInd: 24, color: '#e7b74b'},{xInd: 9, yInd: 25, color: '#e7b74b'},{xInd: 9, yInd: 26, color: '#e7b74b'},{xInd: 9, yInd: 27, color: '#e7b74b'},{xInd: 9, yInd: 28, color: '#f2a557'},{xInd: 10, yInd: 10, color: '#f2a557'},{xInd: 10, yInd: 11, color: '#f2a557'},{xInd: 10, yInd: 12, color: '#f2a557'},{xInd: 10, yInd: 13, color: '#f2a557'},{xInd: 10, yInd: 14, color: '#f2a557'},{xInd: 10, yInd: 15, color: '#f2a557'},{xInd: 10, yInd: 16, color: '#f2a557'},{xInd: 10, yInd: 17, color: '#f2a557'},{xInd: 10, yInd: 18, color: '#f2a557'},{xInd: 10, yInd: 19, color: '#f2a557'},{xInd: 10, yInd: 20, color: '#f2a557'},{xInd: 10, yInd: 21, color: '#f2a557'},{xInd: 10, yInd: 22, color: '#f2a557'},{xInd: 10, yInd: 23, color: '#f2a557'},{xInd: 10, yInd: 24, color: '#f2a557'},{xInd: 10, yInd: 25, color: '#f2a557'},{xInd: 10, yInd: 26, color: '#f2a557'},{xInd: 10, yInd: 27, color: '#f2a557'},{xInd: 10, yInd: 28, color: '#f2a557'},{xInd: 11, yInd: 12, color: '#e7b74b'},{xInd: 11, yInd: 13, color: '#e7b74b'},{xInd: 11, yInd: 14, color: '#e7b74b'},{xInd: 11, yInd: 15, color: '#e7b74b'},{xInd: 11, yInd: 16, color: '#e7b74b'},{xInd: 11, yInd: 17, color: '#e7b74b'},{xInd: 11, yInd: 18, color: '#e7b74b'},{xInd: 11, yInd: 19, color: '#e7b74b'},{xInd: 11, yInd: 20, color: '#e7b74b'},{xInd: 11, yInd: 21, color: '#e7b74b'},{xInd: 11, yInd: 22, color: '#e7b74b'},{xInd: 11, yInd: 23, color: '#e7b74b'},{xInd: 11, yInd: 24, color: '#e7b74b'},{xInd: 11, yInd: 25, color: '#e7b74b'},{xInd: 11, yInd: 26, color: '#e7b74b'},{xInd: 11, yInd: 27, color: '#e7b74b'},{xInd: 11, yInd: 28, color: '#f2a557'},{xInd: 12, yInd: 10, color: '#f2a557'},{xInd: 12, yInd: 11, color: '#f2a557'},{xInd: 12, yInd: 12, color: '#f2a557'},{xInd: 12, yInd: 13, color: '#f2a557'},{xInd: 12, yInd: 14, color: '#f2a557'},{xInd: 12, yInd: 15, color: '#f2a557'},{xInd: 12, yInd: 16, color: '#f2a557'},{xInd: 12, yInd: 17, color: '#f2a557'},{xInd: 12, yInd: 18, color: '#f2a557'},{xInd: 12, yInd: 19, color: '#f2a557'},{xInd: 12, yInd: 20, color: '#f2a557'},{xInd: 12, yInd: 21, color: '#f2a557'},{xInd: 12, yInd: 22, color: '#f2a557'},{xInd: 12, yInd: 23, color: '#f2a557'},{xInd: 12, yInd: 24, color: '#f2a557'},{xInd: 12, yInd: 25, color: '#f2a557'},{xInd: 12, yInd: 26, color: '#f2a557'},{xInd: 12, yInd: 27, color: '#f2a557'},{xInd: 12, yInd: 28, color: '#f2a557'},{xInd: 13, yInd: 12, color: '#e7b74b'},{xInd: 13, yInd: 13, color: '#e7b74b'},{xInd: 13, yInd: 14, color: '#e7b74b'},{xInd: 13, yInd: 15, color: '#e7b74b'},{xInd: 13, yInd: 16, color: '#e7b74b'},{xInd: 13, yInd: 17, color: '#e7b74b'},{xInd: 13, yInd: 18, color: '#e7b74b'},{xInd: 13, yInd: 19, color: '#e7b74b'},{xInd: 13, yInd: 20, color: '#e7b74b'},{xInd: 13, yInd: 21, color: '#e7b74b'},{xInd: 13, yInd: 22, color: '#e7b74b'},{xInd: 13, yInd: 23, color: '#e7b74b'},{xInd: 13, yInd: 24, color: '#e7b74b'},{xInd: 13, yInd: 25, color: '#e7b74b'},{xInd: 13, yInd: 26, color: '#e7b74b'},{xInd: 13, yInd: 27, color: '#e7b74b'},{xInd: 13, yInd: 28, color: '#f2a557'},{xInd: 14, yInd: 10, color: '#f2a557'},{xInd: 14, yInd: 11, color: '#f2a557'},{xInd: 14, yInd: 12, color: '#f2a557'},{xInd: 14, yInd: 13, color: '#f2a557'},{xInd: 14, yInd: 14, color: '#f2a557'},{xInd: 14, yInd: 15, color: '#f2a557'},{xInd: 14, yInd: 16, color: '#f2a557'},{xInd: 14, yInd: 17, color: '#f2a557'},{xInd: 14, yInd: 18, color: '#f2a557'},{xInd: 14, yInd: 19, color: '#f2a557'},{xInd: 14, yInd: 20, color: '#f2a557'},{xInd: 14, yInd: 21, color: '#f2a557'},{xInd: 14, yInd: 22, color: '#f2a557'},{xInd: 14, yInd: 23, color: '#f2a557'},{xInd: 14, yInd: 24, color: '#f2a557'},{xInd: 14, yInd: 25, color: '#f2a557'},{xInd: 14, yInd: 26, color: '#f2a557'},{xInd: 14, yInd: 27, color: '#f2a557'},{xInd: 14, yInd: 28, color: '#f2a557'},{xInd: 15, yInd: 12, color: '#e7b74b'},{xInd: 15, yInd: 13, color: '#e7b74b'},{xInd: 15, yInd: 14, color: '#e7b74b'},{xInd: 15, yInd: 15, color: '#e7b74b'},{xInd: 15, yInd: 16, color: '#e7b74b'},{xInd: 15, yInd: 17, color: '#e7b74b'},{xInd: 15, yInd: 18, color: '#e7b74b'},{xInd: 15, yInd: 19, color: '#e7b74b'},{xInd: 15, yInd: 20, color: '#e7b74b'},{xInd: 15, yInd: 21, color: '#e7b74b'},{xInd: 15, yInd: 22, color: '#e7b74b'},{xInd: 15, yInd: 23, color: '#e7b74b'},{xInd: 15, yInd: 24, color: '#e7b74b'},{xInd: 15, yInd: 25, color: '#e7b74b'},{xInd: 15, yInd: 26, color: '#e7b74b'},{xInd: 15, yInd: 27, color: '#e7b74b'},{xInd: 15, yInd: 28, color: '#f2a557'},{xInd: 16, yInd: 10, color: '#f2a557'},{xInd: 16, yInd: 11, color: '#f2a557'},{xInd: 16, yInd: 12, color: '#f2a557'},{xInd: 16, yInd: 13, color: '#f2a557'},{xInd: 16, yInd: 14, color: '#f2a557'},{xInd: 16, yInd: 15, color: '#f2a557'},{xInd: 16, yInd: 16, color: '#f2a557'},{xInd: 16, yInd: 17, color: '#f2a557'},{xInd: 16, yInd: 18, color: '#f2a557'},{xInd: 16, yInd: 19, color: '#f2a557'},{xInd: 16, yInd: 20, color: '#f2a557'},{xInd: 16, yInd: 21, color: '#f2a557'},{xInd: 16, yInd: 22, color: '#f2a557'},{xInd: 16, yInd: 23, color: '#f2a557'},{xInd: 16, yInd: 24, color: '#f2a557'},{xInd: 16, yInd: 25, color: '#f2a557'},{xInd: 16, yInd: 26, color: '#f2a557'},{xInd: 16, yInd: 27, color: '#f2a557'},{xInd: 16, yInd: 28, color: '#f2a557'},{xInd: 17, yInd: 12, color: '#e7b74b'},{xInd: 17, yInd: 13, color: '#e7b74b'},{xInd: 17, yInd: 14, color: '#e7b74b'},{xInd: 17, yInd: 15, color: '#e7b74b'},{xInd: 17, yInd: 16, color: '#e7b74b'},{xInd: 17, yInd: 17, color: '#e7b74b'},{xInd: 17, yInd: 18, color: '#e7b74b'},{xInd: 17, yInd: 19, color: '#e7b74b'},{xInd: 17, yInd: 20, color: '#e7b74b'},{xInd: 17, yInd: 21, color: '#e7b74b'},{xInd: 17, yInd: 22, color: '#e7b74b'},{xInd: 17, yInd: 23, color: '#e7b74b'},{xInd: 17, yInd: 24, color: '#e7b74b'},{xInd: 17, yInd: 25, color: '#e7b74b'},{xInd: 17, yInd: 26, color: '#e7b74b'},{xInd: 17, yInd: 27, color: '#e7b74b'},{xInd: 17, yInd: 28, color: '#f2a557'},{xInd: 18, yInd: 10, color: '#f2a557'},{xInd: 18, yInd: 11, color: '#f2a557'},{xInd: 18, yInd: 12, color: '#f2a557'},{xInd: 18, yInd: 13, color: '#f2a557'},{xInd: 18, yInd: 14, color: '#f2a557'},{xInd: 18, yInd: 15, color: '#f2a557'},{xInd: 18, yInd: 16, color: '#f2a557'},{xInd: 18, yInd: 17, color: '#f2a557'},{xInd: 18, yInd: 18, color: '#f2a557'},{xInd: 18, yInd: 19, color: '#f2a557'},{xInd: 18, yInd: 20, color: '#f2a557'},{xInd: 18, yInd: 21, color: '#f2a557'},{xInd: 18, yInd: 22, color: '#f2a557'},{xInd: 18, yInd: 23, color: '#f2a557'},{xInd: 18, yInd: 24, color: '#f2a557'},{xInd: 18, yInd: 25, color: '#f2a557'},{xInd: 18, yInd: 26, color: '#f2a557'},{xInd: 18, yInd: 27, color: '#f2a557'},{xInd: 18, yInd: 28, color: '#f2a557'},],
-12: [{xInd: 4, yInd: 4, color: '#e8cfc8'},{xInd: 4, yInd: 5, color: '#e8cfc8'},{xInd: 4, yInd: 6, color: '#e8cfc8'},{xInd: 4, yInd: 7, color: '#e8cfc8'},{xInd: 4, yInd: 8, color: '#e8cfc8'},{xInd: 4, yInd: 9, color: '#e8cfc8'},{xInd: 4, yInd: 10, color: '#e8cfc8'},{xInd: 4, yInd: 11, color: '#e8cfc8'},{xInd: 4, yInd: 12, color: '#e8cfc8'},{xInd: 4, yInd: 13, color: '#e8cfc8'},{xInd: 4, yInd: 14, color: '#e8cfc8'},{xInd: 4, yInd: 15, color: '#e8cfc8'},{xInd: 4, yInd: 16, color: '#e8cfc8'},{xInd: 5, yInd: 4, color: '#e8cfc8'},{xInd: 5, yInd: 5, color: '#e8cfc8'},{xInd: 5, yInd: 6, color: '#e8cfc8'},{xInd: 5, yInd: 7, color: '#e8cfc8'},{xInd: 5, yInd: 8, color: '#e8cfc8'},{xInd: 5, yInd: 9, color: '#e8cfc8'},{xInd: 5, yInd: 10, color: '#e8cfc8'},{xInd: 5, yInd: 13, color: '#e8cfc8'},{xInd: 5, yInd: 14, color: '#e8cfc8'},{xInd: 5, yInd: 15, color: '#e8cfc8'},{xInd: 5, yInd: 16, color: '#e8cfc8'},{xInd: 5, yInd: 17, color: '#e7b74b'},{xInd: 5, yInd: 18, color: '#e7b74b'},{xInd: 5, yInd: 19, color: '#e7b74b'},{xInd: 5, yInd: 22, color: '#e7b74b'},{xInd: 5, yInd: 23, color: '#e7b74b'},{xInd: 5, yInd: 24, color: '#e8cfc8'},{xInd: 5, yInd: 25, color: '#e8cfc8'},{xInd: 6, yInd: 4, color: '#e8cfc8'},{xInd: 6, yInd: 5, color: '#e8cfc8'},{xInd: 6, yInd: 6, color: '#e8cfc8'},{xInd: 6, yInd: 7, color: '#e8cfc8'},{xInd: 6, yInd: 8, color: '#e8cfc8'},{xInd: 6, yInd: 9, color: '#e8cfc8'},{xInd: 6, yInd: 10, color: '#e8cfc8'},{xInd: 6, yInd: 12, color: '#e68b86'},{xInd: 6, yInd: 13, color: '#e8cfc8'},{xInd: 6, yInd: 14, color: '#e8cfc8'},{xInd: 6, yInd: 15, color: '#e8cfc8'},{xInd: 6, yInd: 16, color: '#e8cfc8'},{xInd: 6, yInd: 17, color: '#e8cfc8'},{xInd: 6, yInd: 25, color: '#e8cfc8'},{xInd: 6, yInd: 26, color: '#e8cfc8'},{xInd: 7, yInd: 4, color: '#e8cfc8'},{xInd: 7, yInd: 5, color: '#e8cfc8'},{xInd: 7, yInd: 6, color: '#e8cfc8'},{xInd: 7, yInd: 7, color: '#e8cfc8'},{xInd: 7, yInd: 8, color: '#e8cfc8'},{xInd: 7, yInd: 9, color: '#e8cfc8'},{xInd: 7, yInd: 10, color: '#e8cfc8'},{xInd: 7, yInd: 13, color: '#e8cfc8'},{xInd: 7, yInd: 14, color: '#e8cfc8'},{xInd: 7, yInd: 15, color: '#e8cfc8'},{xInd: 7, yInd: 16, color: '#e8cfc8'},{xInd: 7, yInd: 17, color: '#e8cfc8'},{xInd: 7, yInd: 18, color: '#e8cfc8'},{xInd: 7, yInd: 19, color: '#e7b74b'},{xInd: 7, yInd: 20, color: '#e7b74b'},{xInd: 7, yInd: 21, color: '#e7b74b'},{xInd: 7, yInd: 24, color: '#e7b74b'},{xInd: 7, yInd: 25, color: '#e7b74b'},{xInd: 7, yInd: 26, color: '#e8cfc8'},{xInd: 7, yInd: 27, color: '#e8cfc8'},{xInd: 8, yInd: 4, color: '#e8cfc8'},{xInd: 8, yInd: 5, color: '#e8cfc8'},{xInd: 8, yInd: 6, color: '#e8cfc8'},{xInd: 8, yInd: 7, color: '#e8cfc8'},{xInd: 8, yInd: 8, color: '#e8cfc8'},{xInd: 8, yInd: 9, color: '#e8cfc8'},{xInd: 8, yInd: 10, color: '#e8cfc8'},{xInd: 8, yInd: 11, color: '#e8cfc8'},{xInd: 8, yInd: 12, color: '#e8cfc8'},{xInd: 8, yInd: 13, color: '#e8cfc8'},{xInd: 8, yInd: 14, color: '#e8cfc8'},{xInd: 8, yInd: 15, color: '#e8cfc8'},{xInd: 8, yInd: 16, color: '#e8cfc8'},{xInd: 8, yInd: 17, color: '#e8cfc8'},{xInd: 8, yInd: 18, color: '#e8cfc8'},{xInd: 8, yInd: 19, color: '#e8cfc8'},{xInd: 8, yInd: 26, color: '#e8cfc8'},{xInd: 8, yInd: 27, color: '#e8cfc8'},{xInd: 8, yInd: 28, color: '#e8cfc8'},{xInd: 9, yInd: 4, color: '#e8cfc8'},{xInd: 9, yInd: 5, color: '#e8cfc8'},{xInd: 9, yInd: 6, color: '#e8cfc8'},{xInd: 9, yInd: 7, color: '#e8cfc8'},{xInd: 9, yInd: 8, color: '#e8cfc8'},{xInd: 9, yInd: 9, color: '#e8cfc8'},{xInd: 9, yInd: 10, color: '#e8cfc8'},{xInd: 9, yInd: 11, color: '#e8cfc8'},{xInd: 9, yInd: 12, color: '#e8cfc8'},{xInd: 9, yInd: 13, color: '#e8cfc8'},{xInd: 9, yInd: 14, color: '#e8cfc8'},{xInd: 9, yInd: 17, color: '#e8cfc8'},{xInd: 9, yInd: 18, color: '#e8cfc8'},{xInd: 9, yInd: 19, color: '#e8cfc8'},{xInd: 9, yInd: 26, color: '#e8cfc8'},{xInd: 9, yInd: 27, color: '#e8cfc8'},{xInd: 9, yInd: 28, color: '#e8cfc8'},{xInd: 10, yInd: 4, color: '#e8cfc8'},{xInd: 10, yInd: 5, color: '#e8cfc8'},{xInd: 10, yInd: 6, color: '#e8cfc8'},{xInd: 10, yInd: 7, color: '#e8cfc8'},{xInd: 10, yInd: 8, color: '#e8cfc8'},{xInd: 10, yInd: 9, color: '#e8cfc8'},{xInd: 10, yInd: 10, color: '#e8cfc8'},{xInd: 10, yInd: 11, color: '#e8cfc8'},{xInd: 10, yInd: 12, color: '#e8cfc8'},{xInd: 10, yInd: 13, color: '#e8cfc8'},{xInd: 10, yInd: 14, color: '#e8cfc8'},{xInd: 10, yInd: 15, color: '#e8cfc8'},{xInd: 10, yInd: 16, color: '#e8cfc8'},{xInd: 10, yInd: 17, color: '#e8cfc8'},{xInd: 10, yInd: 18, color: '#e8cfc8'},{xInd: 10, yInd: 19, color: '#e8cfc8'},{xInd: 10, yInd: 26, color: '#e8cfc8'},{xInd: 10, yInd: 27, color: '#e8cfc8'},{xInd: 10, yInd: 28, color: '#e8cfc8'},{xInd: 11, yInd: 4, color: '#e8cfc8'},{xInd: 11, yInd: 5, color: '#e8cfc8'},{xInd: 11, yInd: 6, color: '#e8cfc8'},{xInd: 11, yInd: 7, color: '#e8cfc8'},{xInd: 11, yInd: 8, color: '#e8cfc8'},{xInd: 11, yInd: 9, color: '#e8cfc8'},{xInd: 11, yInd: 10, color: '#e8cfc8'},{xInd: 11, yInd: 13, color: '#e8cfc8'},{xInd: 11, yInd: 14, color: '#e8cfc8'},{xInd: 11, yInd: 15, color: '#e8cfc8'},{xInd: 11, yInd: 16, color: '#e8cfc8'},{xInd: 11, yInd: 17, color: '#e8cfc8'},{xInd: 11, yInd: 18, color: '#e8cfc8'},{xInd: 11, yInd: 19, color: '#e7b74b'},{xInd: 11, yInd: 20, color: '#e7b74b'},{xInd: 11, yInd: 21, color: '#e7b74b'},{xInd: 11, yInd: 24, color: '#e7b74b'},{xInd: 11, yInd: 25, color: '#e7b74b'},{xInd: 11, yInd: 26, color: '#e8cfc8'},{xInd: 11, yInd: 27, color: '#e8cfc8'},{xInd: 12, yInd: 4, color: '#e8cfc8'},{xInd: 12, yInd: 5, color: '#e8cfc8'},{xInd: 12, yInd: 6, color: '#e8cfc8'},{xInd: 12, yInd: 7, color: '#e8cfc8'},{xInd: 12, yInd: 8, color: '#e8cfc8'},{xInd: 12, yInd: 9, color: '#e8cfc8'},{xInd: 12, yInd: 10, color: '#e8cfc8'},{xInd: 12, yInd: 12, color: '#e68b86'},{xInd: 12, yInd: 13, color: '#e8cfc8'},{xInd: 12, yInd: 14, color: '#e8cfc8'},{xInd: 12, yInd: 15, color: '#e8cfc8'},{xInd: 12, yInd: 16, color: '#e8cfc8'},{xInd: 12, yInd: 17, color: '#e8cfc8'},{xInd: 12, yInd: 25, color: '#e8cfc8'},{xInd: 12, yInd: 26, color: '#e8cfc8'},{xInd: 13, yInd: 4, color: '#e8cfc8'},{xInd: 13, yInd: 5, color: '#e8cfc8'},{xInd: 13, yInd: 6, color: '#e8cfc8'},{xInd: 13, yInd: 7, color: '#e8cfc8'},{xInd: 13, yInd: 8, color: '#e8cfc8'},{xInd: 13, yInd: 9, color: '#e8cfc8'},{xInd: 13, yInd: 10, color: '#e8cfc8'},{xInd: 13, yInd: 13, color: '#e8cfc8'},{xInd: 13, yInd: 14, color: '#e8cfc8'},{xInd: 13, yInd: 15, color: '#e8cfc8'},{xInd: 13, yInd: 16, color: '#e8cfc8'},{xInd: 13, yInd: 17, color: '#e7b74b'},{xInd: 13, yInd: 18, color: '#e7b74b'},{xInd: 13, yInd: 19, color: '#e7b74b'},{xInd: 13, yInd: 22, color: '#e7b74b'},{xInd: 13, yInd: 23, color: '#e7b74b'},{xInd: 13, yInd: 24, color: '#e8cfc8'},{xInd: 13, yInd: 25, color: '#e8cfc8'},{xInd: 14, yInd: 4, color: '#e8cfc8'},{xInd: 14, yInd: 5, color: '#e8cfc8'},{xInd: 14, yInd: 6, color: '#e8cfc8'},{xInd: 14, yInd: 7, color: '#e8cfc8'},{xInd: 14, yInd: 8, color: '#e8cfc8'},{xInd: 14, yInd: 9, color: '#e8cfc8'},{xInd: 14, yInd: 10, color: '#e8cfc8'},{xInd: 14, yInd: 11, color: '#e8cfc8'},{xInd: 14, yInd: 12, color: '#e8cfc8'},{xInd: 14, yInd: 13, color: '#e8cfc8'},{xInd: 14, yInd: 14, color: '#e8cfc8'},{xInd: 14, yInd: 15, color: '#e8cfc8'},{xInd: 14, yInd: 16, color: '#e8cfc8'},],
-
-}
