+++ src/stage.js
