+++ test/drawBall.test.js
@@ -1,6 +1,6 @@
-import { drawBall } from "../drawBall.js";
-import { Ball } from "../ball.js";
-import { Paddle } from "../paddle.js";
+import { drawBall } from "../src/drawBall.js";
+import { Ball } from "../src/ball.js";
+import { Paddle } from "../src/paddle.js";
 
 const cvs = { width: 400, height: 200 };
 const ctx = {
