--- v2/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Germania+One">
-    <link rel="stylesheet" href="css/style.css">
-	<title>My Game</title>
-</head>
-<body>
-	<div id="msg"></div>
-
-	<div class="sound">
-        <img src="img/SOUND_ON.png" alt="" id="sound">
-    </div>
-    
-    <div id="gameover">
-        <img src="img/youwon.png" alt="" id="youwon">
-        <img src="img/gameover.png" alt="" id="youlose">
-        <div id="restart">Play Again!</div>
-    </div>
-
-	<canvas id="breakout" width="400" height="500"></canvas>
-	<script type="module" src="script.js" ></script>	
-</body>
-</html>
